Execution contexts

Jaynes.GenerateContext — Type

mutable struct GenerateContext{T <: Trace, K <: ConstrainedSelection} <: ExecutionContext
    tr::T
    select::K
    weight::Float64
    visited::Visitor
    params::LearnableParameters
end

GenerateContext is used to generate traces, as well as record and accumulate likelihood weights given observations at addressed randomness.

Inner constructors:

GenerateContext(tr::T, select::K) where {T <: Trace, K <: ConstrainedSelection} = new{T, K}(tr, select, 0.0, Visitor(), LearnableParameters())
GenerateContext(tr::T, select::K, params) where {T <: Trace, K <: ConstrainedSelection} = new{T, K}(tr, select, 0.0, Visitor(), params)

Outer constructs:

Generate(select::ConstrainedSelection) = GenerateContext(Trace(), select)
Generate(select::ConstrainedSelection, params) = GenerateContext(Trace(), select, params)
Generate(tr::Trace, select::ConstrainedSelection) = GenerateContext(tr, select)

Jaynes.ProposeContext — Type

mutable struct ProposeContext{T <: Trace} <: ExecutionContext
    tr::T
    weight::Float64
    params::LearnableParameters
end

ProposeContext is used to generate traces for inference algorithms which use custom proposals. ProposeContext instances can be passed sets of LearnableParameters to configure the propose with parameters which have been learned by differentiable programming.

Inner constructors:

ProposeContext(tr::T) where T <: Trace = new{T}(tr, 0.0, LearnableParameters())

Outer constructs:

Propose() = ProposeContext(Trace())

Missing docstring.

Missing docstring for ScoreContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UpdateContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RegenerateContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BackpropagateContext. Check Documenter's build log for details.