Jaynes supports Zygote-based reverse mode gradient computation of learnable parameters and primitive probabilistic choices. This functionality is accessed through two different gradient contexts.

Jaynes.ParameterBackpropagateContext — Type

mutable struct ParameterBackpropagateContext{T <: Trace} <: BackpropagationContext
    tr::T
    weight::Float64
    visited::Visitor
    params::ParameterStore
    param_grads::Gradients
end

ParameterBackpropagateContext is used to compute the gradients of parameters with respect to following objective:

Outer constructors:

ParameterBackpropagate(tr::T, params) where T <: Trace = ParameterBackpropagateContext(tr, 0.0, Visitor(), params, Gradients())
ParameterBackpropagate(tr::T, params, param_grads::Gradients) where {T <: Trace, K <: UnconstrainedSelection} = ParameterBackpropagateContext(tr, 0.0, Visitor(), params, param_grads)

For one-shot gradient computations, this context is easily accessed through the get_parameter_gradients method.

Jaynes.get_parameter_gradients — Function

gradients = get_parameter_gradients(cl::T, ret_grad, scaler::Float64 = 1.0) where T <: CallSite

Returns a Gradients object which tracks the gradients of the objective with respect to parameters in the program.

For choices, the context is ChoiceBackpropagateContext.

Jaynes.ChoiceBackpropagateContext — Type

mutable struct ChoiceBackpropagateContext{T <: Trace} <: BackpropagationContext
    tr::T
    weight::Float64
    visited::Visitor
    params::ParameterStore
    param_grads::Gradients
end

ChoiceBackpropagateContext is used to compute the gradients of choices with respect to following objective:

Outer constructors:

ChoiceBackpropagate(tr::T, params, choice_grads) where {T <: Trace, K <: UnconstrainedSelection} = ChoiceBackpropagateContext(tr, 0.0, Visitor(), params, choice_grads, UnconstrainedAllSelection())
ChoiceBackpropagate(tr::T, params, choice_grads, sel::K) where {T <: Trace, K <: UnconstrainedSelection} = ChoiceBackpropagateContext(tr, 0.0, Visitor(), params, choice_grads, sel)

For one-shot gradient computations on choices, the ChoiceBackpropagateContext is easily accessed through the get_choice_gradients method.

Jaynes.get_choice_gradients — Function

gradients = get_choice_gradients(cl::T, ret_grad) where T <: CallSite

Returns a Gradients object which tracks the gradients with respect to the objective of random choices with differentiable logpdf in the program.

In the future, Jaynes will support a context which allows the automatic training of neural network components (Flux.jl or otherwise) facilitated by custom call sites. See the foreign model interface for more details.

As part of standard usage, itt's likely that you'd like to update learnable parameters in your model (which you declare with learnable(addr, initial_value). Given a CallSite, you can extract parameters using get_parameters

ret, cl, w = generate(selection((:q, -0.5)), learnable_hypers)
params = get_parameters(cl)

which produces an instance of LearnableParameters by unpacking the params field in any trace kept in the call site. LearnableParameters and Gradients are explicitly kept separate from the internal call site representation, to encourage portability, as well as non-standard optimization schemes. Given a Gradients instance, to update your parameters, just call update!

update!(params, gradients)

which mutates the parameters in-place. You can then pass these in as an argument to the normal context interfaces (e.g. generate, simulate, etc) to use your updated parameters.