Jaynes.initialize_filter — Function

particles = initialize_filter(observations::ConstrainedHierarchicalSelection,
                              num_particles::Int,
                              fn::Function, 
                              args::Tuple)

Instantiate a set of particles using a call to importance_sampling.

Jaynes.filter_step! — Function

filter_step!(observations::ConstrainedHierarchicalSelection,
             ps::Particles,
             new_args::Tuple)

Perform a single filter step from an instance ps of Particles, applying the constraints specified by observations.

filter_step!(observations::ConstrainedHierarchicalSelection,
             ps::Particles,
             new_args::Tuple,
             proposal::Function,
             proposal_args::Tuple)

Perform a single filter step using a custom proposal function, applying the constraints specified by observations.

Info

Custom proposals provided to filter_step! should accept as first argument a CallSite instance (e.g. so that you can use the previous trace and information in your transition proposal).

Jaynes.check_ess_resample! — Function

check_ess_resample!(ps::Particles)

Checks the effective sample size using ess, then resamples from an existing instance of Particles by mutation in place.