Here is a description of the set of "standard" contexts which are used frequently in modeling and inference. Gradient and foreign model contexts are discussed in Differentiable programming and Foreign model interface, respectively.

Jaynes.GenerateContext — Type

mutable struct GenerateContext{T <: AddressMap, K <: AddressMap, P <: Parameters} <: ExecutionContext
     tr::T
     target::K
     weight::Float64
     score::Float64
     visited::Visitor
     params::P
end

GenerateContext is used to generate traces, as well as record and accumulate likelihood weights given observations at addressed randomness.

Inner constructors:

GenerateContext(tr::T, target::K) where {T <: AddressMap, K <: AddressMap} = new{T, K}(tr, target, 0.0, Visitor(), Parameters())
GenerateContext(tr::T, target::K, params::P) where {T <: AddressMap, K <: AddressMap, P <: Parameters} = new{T, K, P}(tr, target, 0.0, Visitor(), params)

Outer constructors:

Generate(target::AddressMap) = GenerateContext(AddressMap(), target)
Generate(target::AddressMap, params) = GenerateContext(AddressMap(), target, params)
Generate(tr::AddressMap, target::AddressMap) = GenerateContext(tr, target)

Jaynes.generate — Function

ret, cl, w = generate(target::L, fn::Function, args...; params = Parameters()) where L <: AddressMap
ret, cs, w = generate(target::L, fn::typeof(rand), d::Distribution{K}; params = Parameters()) where {L <: AddressMap, K}
ret, v_cl, w = generate(target::L, fn::typeof(markov), call::Function, len::Int, args...; params = Parameters()) where L <: AddressMap
ret, v_cl, w = generate(target::L, fn::typeof(plate), call::Function, args::Vector; params = Parameters()) where L <: AddressMap
ret, v_cl, w = generate(target::L, fn::typeof(plate), d::Distribution{K}, len::Int; params = Parameters()) where {L <: AddressMap, K}

generate provides an API to the GenerateContext execution context. You can use this function on any of the matching signatures above - it will return the return value ret, a RecordSite instance specialized to the call, and the score/weight w computed with respect to the constraints target.

Jaynes.SimulateContext — Type

mutable struct SimulateContext{T <: AddressMap, P <: AddressMap} <: ExecutionContext
    tr::T
    visited::Visitor
    params::P
    SimulateContext(params) where T <: AddressMap = new{T}(AddressMap(), Visitor(), params)
end

SimulateContext is used to simulate traces without recording likelihood weights. SimulateContext can be instantiated with custom AddressMap instances, which is useful when used for gradient-based learning.

Inner constructors:

SimulateContext(params) = new{DynamicAddressMap}(AddressMap(), Visitor(), params)

Jaynes.simulate — Function

ret, cl = simulate(fn::Function, args...; params = LearnableByAddress())
ret, cl = simulate(fn::typeof(rand), d::Distribution{T}; params = LearnableByAddress()) where T
ret, v_cl = simulate(c::typeof(plate), fn::Function, args::Vector; params = LearnableByAddress()) where T
ret, v_cl = simulate(fn::typeof(plate), d::Distribution{T}, len::Int; params = LearnableByAddress()) where T
ret, v_cl = simulate(c::typeof(markov), fn::Function, len::Int, args...; params = LearnableByAddress())

simulate function provides an API to the SimulateContext execution context. You can use this function on any of the matching signatures above - it will return the return value ret, and a RecordSite instance specialized to the call. simulate is used to express unconstrained generation of a probabilistic program trace, without likelihood weight recording.

Jaynes.ProposeContext — Type

mutable struct ProposeContext{T <: AddressMap, P <: AddressMap} <: ExecutionContext
    tr::T
    score::Float64
    visited::Visitor
    params::P
end

ProposeContext is used to propose traces for inference algorithms which use custom proposals. ProposeContext instances can be passed sets of AddressMap to configure the propose with parameters which have been learned by differentiable programming.

Inner constructors:

ProposeContext(tr::T) where T <: AddressMap = new{T}(tr, 0.0, AddressMap())

Outer constructors:

Propose() = ProposeContext(AddressMap())

Jaynes.propose — Function

ret, g_cl, w = propose(fn::Function, args...)
ret, cs, w = propose(fn::typeof(rand), d::Distribution{K}) where K
ret, v_cl, w = propose(fn::typeof(markov), call::Function, len::Int, args...)
ret, v_cl, w = propose(fn::typeof(plate), call::Function, args::Vector)
ret, v_cl, w = propose(fn::typeof(plate), d::Distribution{K}, len::Int) where K

propose provides an API to the ProposeContext execution context. You can use this function on any of the matching signatures above - it will return the return value ret, a RecordSite instance specialized to the call, and the score w.

Jaynes.ScoreContext — Type

mutable struct ScoreContext{P <: AddressMap} <: ExecutionContext
    select::AddressMap
    weight::Float64
    params::P
end

The ScoreContext is used to score selections according to a model function. For computation in the ScoreContext to execute successfully, the select selection must provide a choice for every address visited in the model function, and the model function must allow the context to visit every constraints expressed in select.

Inner constructors:

function Score(obs::Vector{Tuple{K, P}}) where {P, K <: Union{Symbol, Pair}}
    c_sel = selection(obs)
    new{EmptyAddressMap}(c_sel, 0.0, AddressMap())
end

Outer constructors:

ScoreContext(obs::K, params) where {K <: AddressMap} = new(obs, 0.0, params)
Score(obs::Vector) = ScoreContext(selection(obs))
Score(obs::AddressMap) = ScoreContext(obs, AddressMap())
Score(obs::AddressMap, params) = ScoreContext(obs, params)

Jaynes.score — Function

ret, w = score(sel::L, fn::Function, args...; params = AddressMap()) where L <: AddressMap
ret, w = score(sel::L, fn::typeof(rand), d::Distribution{K}; params = AddressMap()) where {L <: AddressMap, K}
ret, w = score(sel::L, fn::typeof(markov), call::Function, len::Int, args...; params = AddressMap()) where L <: AddressMap
ret, w = score(sel::L, fn::typeof(plate), call::Function, args::Vector; params = AddressMap()) where L <: AddressMap
ret, w = score(sel::L, fn::typeof(plate), d::Distribution{K}, len::Int; params = AddressMap()) where {L <: AddressMap, K}

score provides an API to the ScoreContext execution context. You can use this function on any of the matching signatures above - it will return the return value ret, and the likelihood weight score of the user-provided selection sel. The selection should satisfy the following requirement:

At any random choice in any branch traveled according to the constraints of sel, sel must provide a constraint for that choice.

Simply put, this just means you need to provide a constraint for each ChoiceSite you encounter.

Jaynes.UpdateContext — Type

mutable struct UpdateContext{C <: CallSite, 
                             T <: AddressMap,
                             K <: AddressMap, 
                             P <: AddressMap, 
                             D <: Diff} <: ExecutionContext
    prev::C
    tr::T
    select::K
    weight::Float64
    score::Float64
    discard::DynamicAddressMap
    visited::Visitor
    params::P
    argdiffs::D
end

Inner constructor:

UpdateContext(cl::C, select::K, argdiffs::D) where {C <: CallSite, K <: AddressMap, D <: Diff} = new{C, typeof(cl.trace), K, EmptyAddressMap, D}(cl, typeof(cl.trace)(), select, 0.0, 0.0, AddressMap(), Visitor(), AddressMap(), argdiffs)
UpdateContext(cl::C, select::K, ps::P, argdiffs::D) where {C <: CallSite, K <: AddressMap, P <: AddressMap, D <: Diff} = new{C, typeof(cl.trace), K, EmptyAddressMap, D}(cl, typeof(cl.trace)(), select, 0.0, 0.0, AddressMap(), Visitor(), ps, argdiffs)

UpdateContext is an execution context used for updating the value of random choices in an existing recorded call site. This context will perform corrective updates to the likehood weights and scores so that this operation produces the correct weights and scores for the original model program constrained with the select selection in the UpdateContext.

Jaynes.update — Function

ret, cl, w, retdiff, d = update(ctx::UpdateContext, cs::DynamicCallSite, args...) where D <: Diff
ret, cl, w, retdiff, d = update(sel::L, cs::DynamicCallSite) where L <: AddressMap
ret, cl, w, retdiff, d = update(sel::L, cs::DynamicCallSite, argdiffs::D, new_args...) where {L <: AddressMap, D <: Diff}
ret, v_cl, w, retdiff, d = update(sel::L, vcs::VectorizedCallSite{typeof(plate)}) where {L <: AddressMap, D <: Diff}
ret, v_cl, w, retdiff, d = update(sel::L, vcs::VectorizedCallSite{typeof(markov)}) where {L <: AddressMap, D <: Diff}
ret, v_cl, w, retdiff, d = update(sel::L, vcs::VectorizedCallSite{typeof(markov)}, d::NoChange, len::Int) where {L <: AddressMap, D <: Diff}
ret, v_cl, w, retdiff, d = update(sel::L, vcs::VectorizedCallSite{typeof(markov)}, len::Int) where {L <: AddressMap, D <: Diff}

update provides an API to the UpdateContext execution context. You can use this function on any of the matching signatures above - it will return the return value ret, the updated RecordSite instance cl or v_cl, the updated weight w, a Diff instance for the return value retdiff, and a structure which contains any changed (i.e. discarded) record sites d.

Jaynes.RegenerateContext — Type

mutable struct RegenerateContext{T <: Trace, 
                                 L <: Target,
                                 P <: AddressMap} <: ExecutionContext
    prev::T
    tr::T
    target::L
    weight::Float64
    score::Float64
    discard::T
    visited::Visitor
    params::P
end

Inner constructors:

function RegenerateContext(tr::T, sel::Vector{Address}) where T <: Trace
    un_sel = targetion(sel)
    new{T, typeof(un_sel), EmptyAddressMap}(tr, Trace(), un_sel, 0.0, Trace(), Visitor(), AddressMap())
end
function RegenerateContext(tr::T, sel::L) where {T <: Trace, L <: Target}
    new{T, L, EmptyAddressMap}(tr, Trace(), sel, 0.0, Trace(), Visitor(), AddressMap())
end

Outer constructors:

Regenerate(tr::Trace, sel::Vector{Address}) = RegenerateContext(tr, sel)
Regenerate(tr::Trace, sel::Target) = RegenerateContext(tr, sel)

The RegenerateContext is used for MCMC algorithms, to propose new choices for addresses indicated by an Target in the target field.

Jaynes.regenerate — Function

ret, cl = regenerate(sel::L, cs::DynamicCallSite, new_args...) where L <: Target
ret, cl = regenerate(sel::L, cs::DynamicCallSite) where L <: Target

regenerate is an API to the RegenerateContext execution context. regenerate requires that users provide an Target, an original call site, and possibly a set of new arguments to be used in the regeneration step. This context internally keeps track of the bookkeeping required to increment likelihood weights, as well as prune off parts of the trace which are invalid if a regenerated choice changes the shape of the trace (e.g. control flow), and returns a new return value ret as well as the modified call site cl.