Due to the design and implementation as an IR metaprogramming tool, Jaynes sits at a slightly privileged place in the probabilistic programming ecosystem, in the sense that many of the other languages which users are likely to use require the usage of macros to setup code in a way which allows the necessary state to be inserted for probabilistic programming functionality.

Jaynes sees all the code after macro expansion is completed, which allows Jaynes to introspect function call sites after state has been inserted by other libraries. This allows the possibility for Jaynes to construct special call sites to represent calls into other probabilistic programming libraries. These interfaces are a work in progress, but Jaynes should theoretically provide a lingua franca for programs expressed in different probabilistic programming systems to communicate in a natural way, due to the nature of the context-oriented programming style facilitated by the system.

Black-box extensions

Jaynes.primitive — Function

@primitive function logpdf(fn::typeof(foo), args, foo_ret)
    ...
end

@primitive is a convenience metaprogramming construct which derives contextual dispatch definitions for functions which you'd like the tracer to "summarize" to a single site. This is one mechanism which gives the user more control over the structure of the choice map structure in their programs.

Example:

geo(p::Float64) = rand(:flip, Bernoulli(p)) ? 1 : 1 + rand(:geo, geo, p)

# Define as primitive.
@primitive function logpdf(fn::typeof(geo), p, count)
    return Distributions.logpdf(Geometric(p), count)
end

function foo()
    ret = rand(:geo, geo, 0.3)
    ret
end

The above program would summarize the trace into a single choice site for :geo, as if geo was a primitive distribution.

  __________________________________

               Addresses

 geo : 42
  __________________________________

Foreign models

Jaynes.foreign — Function

foreign(addr::A, m, args...) where A <: Address

Activate a foreign model interface. The tracer will treat this is a specialized call site, depending on the type of m. Currently supports typeof(m) <: Soss.Model and typeof(m) <: Gen.GenerativeFunction.

Jaynes.load_soss_fmi — Function

Jaynes.@load_soss_fmi

@load_soss_fmi loads the Soss.jl foreign model interface extension. This allows you to utilize Soss models in your modeling.

Example:

Jaynes.@load_soss_fmi()

# A Soss model.
m = @model σ begin
    μ ~ Normal()
    y ~ Normal(μ, σ) |> iid(5)
end

bar = () -> begin
    x = rand(:x, Normal(5.0, 1.0))
    soss_ret = foreign(:foo, m, (σ = x,))
    return soss_ret
end

This interface currently supports all the inference interfaces (e.g. simulate, generate, score, regenerate, update, propose) which means that you can use any of the inference algorithms in the standard inference library.

Jaynes.load_gen_fmi — Function

Jaynes.@load_gen_fmi

@load_gen_fmi loads the Gen.jl foreign model interface extension. This allows you to utilize Gen models (in any of Gen's DSLs) in your modeling.

Example:

Jaynes.@load_gen_fmi()

@gen (static) function foo(z::Float64)
    x = @trace(normal(z, 1.0), :x)
    y = @trace(normal(x, 1.0), :y)
    return x
end

Gen.load_generated_functions()

bar = () -> begin
    x = rand(:x, Normal(0.0, 1.0))
    return foreign(:foo, foo, x)
end

ret, cl = Jaynes.simulate(bar)

This interface currently supports all the inference interfaces (e.g. simulate, generate, score, regenerate, update, propose) which means that you can use any of the inference algorithms in the standard inference library.