Importance sampling

Jaynes.importance_sampling — Function

Samples from the model prior.

particles, normalized_weights = importance_sampling(model::Function, 
                                                    args::Tuple; 
                                                    observations::ConstrainedSelection = ConstrainedAnywhereSelection(), 
                                                    num_samples::Int = 5000)

Samples from a programmer-provided proposal function.

particles, normalized_weights = importance_sampling(model::Function, 
                                                    args::Tuple, 
                                                    proposal::Function, 
                                                    proposal_args::Tuple; 
                                                    observations::ConstrainedSelection = ConstrainedAnywhereSelection(), 
                                                    num_samples::Int = 5000)

Run importance sampling on the posterior over unconstrained addresses and values. Returns an instance of Particles and normalized weights.

Info

Addressed randomness in a custom proposal passed to importance_sampling should satisfy the following criteria to ensure that inference is mathematically valid:

Custom proposals should only propose to unobserved addresses in the original program.

Custom proposals should not propose to addresses which do not occur in the original program.

Particle filtering

Jaynes.initialize_filter — Function

particles = initialize_filter(fn::Function, 
                              args::Tuple,
                              observations::ConstrainedHierarchicalSelection,
                              num_particles::Int)

Instantiate a set of particles using a call to importance_sampling.

Jaynes.filter_step! — Function

filter_step!(ps::Particles,
             new_args::Tuple,
             observations::ConstrainedHierarchicalSelection)

Perform a single filter step from an instance ps of Particles, applying the constraints specified by observations.

filter_step!(ps::Particles,
             new_args::Tuple,
             proposal::Function,
             proposal_args::Tuple,
             observations::ConstrainedHierarchicalSelection)

Perform a single filter step using a custom proposal function, applying the constraints specified by observations.

Info

Custom proposals provided to filter_step! should accept as first argument a CallSite instance (e.g. so that you can use the previous trace and information in your transition proposal).

Jaynes.resample! — Function

resample!(ps::Particles)
resample!(ps::Particles, num::Int)

Resample from an existing instance of Particles by mutation in place.

resample! can be applied to both instances of Particles produced by particle filtering, as well as instances of Particles produced by importance sampling.

Markov chain Monte Carlo

Jaynes.metropolis_hastings — Function

call, accepted, metropolis_hastings(call::BlackBoxCallSite,
                                    sel::UnconstrainedSelection)

Perform a Metropolis-Hastings step by proposing new choices using the prior at addressed specified by sel. Returns a call site, as well as a Boolean value accepted to indicate if the proposal was accepted or rejected.

call, accepted = metropolis_hastings(call::BlackBoxCallSite,
                                     proposal::Function,
                                     proposal_args::Tuple,
                                     sel::UnconstrainedSelection)

Perform a Metropolis-Hastings step by proposing new choices using a custom proposal at addressed specified by sel. Returns a call site, as well as a Boolean value accepted to indicate if the proposal was accepted or rejected.

Info

Similar to custom proposals for particle filtering, a custom proposal passed to metropolis_hastings should accept an instance of CallSite as the first argument, followed by the other proposal arguments.

Additionally, if the proposal proposes to addresses which modify the control flow of the original call, it must also provide proposal choice sites for any addressed which exist on that branch of the original model. If this criterion is not satisfied, the kernel is not stationary with respect to the target posterior of the model.