Jaynes.initialize_filter — Function

particles = initialize_filter(fn::Function, 
                              args::Tuple,
                              observations::ConstrainedHierarchicalSelection,
                              num_particles::Int)

Instantiate a set of particles using a call to importance_sampling.

Jaynes.filter_step! — Function

filter_step!(ps::Particles,
             new_args::Tuple,
             observations::ConstrainedHierarchicalSelection)

Perform a single filter step from an instance ps of Particles, applying the constraints specified by observations.

filter_step!(ps::Particles,
             new_args::Tuple,
             proposal::Function,
             proposal_args::Tuple,
             observations::ConstrainedHierarchicalSelection)

Perform a single filter step using a custom proposal function, applying the constraints specified by observations.

Info

Custom proposals provided to filter_step! should accept as first argument a CallSite instance (e.g. so that you can use the previous trace and information in your transition proposal).

Jaynes.resample! — Function

resample!(ps::Particles)
resample!(ps::Particles, num::Int)

Resample from an existing instance of Particles by mutation in place.

resample! can be applied to both instances of Particles produced by particle filtering, as well as instances of Particles produced by importance sampling.