Here is a description of the set of "standard" contexts which are used frequently in modeling and inference. Gradient and foreign model contexts are discussed in Differentiable programming and Foreign model interface, respectively.

Jaynes.GenerateContext — Type

mutable struct GenerateContext{T <: Trace, K <: ConstrainedSelection} <: ExecutionContext
     tr::T
     select::K
     weight::Float64
     visited::Visitor
     params::LearnableParameters
end

GenerateContext is used to generate traces, as well as record and accumulate likelihood weights given observations at addressed randomness.

Inner constructors:

GenerateContext(tr::T, select::K) where {T <: Trace, K <: ConstrainedSelection} = new{T, K}(tr, select, 0.0, Visitor(), LearnableParameters())
GenerateContext(tr::T, select::K, params) where {T <: Trace, K <: ConstrainedSelection} = new{T, K}(tr, select, 0.0, Visitor(), params)

Outer constructors:

Generate(select::ConstrainedSelection) = GenerateContext(Trace(), select)
Generate(select::ConstrainedSelection, params) = GenerateContext(Trace(), select, params)
Generate(tr::Trace, select::ConstrainedSelection) = GenerateContext(tr, select)

Missing docstring.

Missing docstring for generate. Check Documenter's build log for details.

Jaynes.SimulateContext — Type

mutable struct SimulateContext{T <: Trace} <: ExecutionContext
    tr::T
    visited::Visitor
    params::LearnableParameters
    SimulateContext(params) where T <: Trace = new{T}(Trace(), Visitor(), params)
end

SimulateContext is used to simulate traces without recording likelihood weights. SimulateContext can be instantiated with custom LearnableParameters instances, which is useful when used for gradient-based learning.

Inner constructors:

SimulateContext(params) = new{HierarchicalTrace}(Trace(), Visitor(), params)

Missing docstring.

Missing docstring for simulate. Check Documenter's build log for details.

Jaynes.ProposeContext — Type

mutable struct ProposeContext{T <: Trace} <: ExecutionContext
    tr::T
    weight::Float64
    params::LearnableParameters
end

ProposeContext is used to generate traces for inference algorithms which use custom proposals. ProposeContext instances can be passed sets of LearnableParameters to configure the propose with parameters which have been learned by differentiable programming.

Inner constructors:

ProposeContext(tr::T) where T <: Trace = new{T}(tr, 0.0, LearnableParameters())

Outer constructors:

Propose() = ProposeContext(Trace())

Missing docstring.

Missing docstring for propose. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ScoreContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for score. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UpdateContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for update. Check Documenter's build log for details.

Jaynes.RegenerateContext — Type

mutable struct RegenerateContext{T <: Trace, L <: UnconstrainedSelection} <: ExecutionContext
    prev::T
    tr::T
    select::L
    weight::Float64
    score::Float64
    discard::T
    visited::Visitor
    params::LearnableParameters
end

Inner constructors:

function RegenerateContext(tr::T, sel::Vector{Address}) where T <: Trace
    un_sel = selection(sel)
    new{T, typeof(un_sel)}(tr, Trace(), un_sel, 0.0, Trace(), Visitor(), LearnableParameters())
end
function RegenerateContext(tr::T, sel::L) where {T <: Trace, L <: UnconstrainedSelection}
    new{T, L}(tr, Trace(), sel, 0.0, Trace(), Visitor(), LearnableParameters())
end

Outer constructors:

Regenerate(tr::Trace, sel::Vector{Address}) = RegenerateContext(tr, sel)
Regenerate(tr::Trace, sel::UnconstrainedSelection) = RegenerateContext(tr, sel)

The RegenerateContext is used for MCMC algorithms, to propose new choices for addresses indicated by an UnconstrainedSelection in the select field.

Jaynes.regenerate — Function

ret, call_site = regenerate(sel::L, bbcs::BlackBoxCallSite, new_args...) where L <: UnconstrainedSelection
ret, call_site = regenerate(sel::L, bbcs::BlackBoxCallSite) where L <: UnconstrainedSelection

Users will likely interact with the RegenerateContext through the convenience method regenerate which requires that users provide an UnconstrainedSelection, an original call site, and possibly a set of new arguments to be used in the regeneration step. This context internally keeps track of the bookkeeping required to increment likelihood weights, as well as prune off parts of the trace which are invalid if a regenerated choice changes the shape of the trace (e.g. control flow).