Jaynes.Trace — Type

abstract type Trace end

Abstract base type of all traces.

Jaynes.HierarchicalTrace — Type

struct HierarchicalTrace <: Trace
    calls::Dict{Address, CallSite}
    choices::Dict{Address, ChoiceSite}
    params::Dict{Address, LearnableSite}
    end
end

Execution trace which allows the tracking of randomness metadata in a function call.

Jaynes.VectorizedTrace — Type

struct VectorizedTrace{C <: RecordSite} <: Trace
    subrecords::Vector{C}
    params::Dict{Address, LearnableSite}
end

Structured execution trace for markov and plate calls. The dependency structure interpretation of the subrecords vector depends on the call. For markov, the structure is Markovian. For plate, each element is drawn IID from the program or distribution provided to plate.

Jaynes.ChoiceSite — Type

struct ChoiceSite{T} <: RecordSite
    score::Float64
    val::T
end

A record of a random sample at an addressed rand call with a Distribution instance. Keeps the value of the random sample and the logpdf score.

Jaynes.HierarchicalCallSite — Type

struct HierarchicalCallSite{J, K} <: CallSite
    trace::HierarchicalTrace
    score::Float64
    fn::Function
    args::J
    ret::K
end

A record of a black-box call (e.g. no special tracer language features). Records the fn and args for the call, as well as the ret return value.

Jaynes.VectorizedCallSite — Type

struct VectorizedCallSite{F <: Function, C <: RecordSite, J, K} <: CallSite
    trace::VectorizedTrace{C}
    score::Float64
    fn::Function
    args::J
    ret::Vector{K}
end

A record of a call site using the special plate and markov tracer language features. Informs the tracer that the call conforms to a special pattern of randomness dependency, which allows the storing of Trace instances sequentially in a vector.