Jaynes.Trace — Type

abstract type Trace end

Abstract base type of all traces.

Jaynes.HierarchicalTrace — Type

mutable struct HierarchicalTrace <: Trace
    calls::Dict{Address, CallSite}
    choices::Dict{Address, ChoiceSite}
    params::Dict{Address, LearnableSite}
    score::Float64
    end
end

Structured execution trace with tracked randomness in a function call.

Jaynes.ChoiceSite — Type

struct ChoiceSite{T} <: RecordSite
    score::Float64
    val::T
end

A record of a random sample at an addressed rand call with a Distribution instance. Keeps the value of the random sample and the logpdf score.

Jaynes.BlackBoxCallSite — Type

mutable struct BlackBoxCallSite{T <: Trace, J, K} <: CallSite
    trace::T
    fn::Function
    args::J
    ret::K
end

A record of a black-box call (e.g. no special tracer language features). Records the fn and args for the call, as well as the ret return value.

Jaynes.VectorizedCallSite — Type

mutable struct VectorizedCallSite{F <: Function, C <: RecordSite, J, K} <: CallSite
    subcalls::Vector{T}
    score::Float64
    fn::Function
    args::J
    ret::Vector{K}
end

A record of a call site using the special plate and markov tracer language features. Informs the tracer that the call conforms to a special pattern of randomness dependency, which allows the storing of Trace instances sequentially in a vector.