Here is a description of the set of "standard" contexts which are used frequently in modeling and inference. Gradient and foreign model contexts are discussed in Differentiable programming and Foreign model interface, respectively.

Jaynes.GenerateContext — Type

mutable struct GenerateContext{T <: Trace, K <: ConstrainedSelection, P <: Parameters} <: ExecutionContext
     tr::T
     select::K
     weight::Float64
     score::Float64
     visited::Visitor
     params::P
end

GenerateContext is used to generate traces, as well as record and accumulate likelihood weights given observations at addressed randomness.

Inner constructors:

GenerateContext(tr::T, select::K) where {T <: Trace, K <: ConstrainedSelection} = new{T, K}(tr, select, 0.0, Visitor(), Parameters())
GenerateContext(tr::T, select::K, params::P) where {T <: Trace, K <: ConstrainedSelection, P <: Parameters} = new{T, K, P}(tr, select, 0.0, Visitor(), params)

Outer constructors:

Generate(select::ConstrainedSelection) = GenerateContext(Trace(), select)
Generate(select::ConstrainedSelection, params) = GenerateContext(Trace(), select, params)
Generate(tr::Trace, select::ConstrainedSelection) = GenerateContext(tr, select)

Jaynes.generate — Function

ret, cl, w = generate(sel::L, fn::Function, args...; params = Parameters()) where L <: ConstrainedSelection
ret, cs, w = generate(sel::L, fn::typeof(rand), d::Distribution{K}; params = Parameters()) where {L <: ConstrainedSelection, K}
ret, v_cl, w = generate(sel::L, fn::typeof(markov), call::Function, len::Int, args...; params = Parameters()) where L <: ConstrainedSelection
ret, v_cl, w = generate(sel::L, fn::typeof(plate), call::Function, args::Vector; params = Parameters()) where L <: ConstrainedSelection
ret, v_cl, w = generate(sel::L, fn::typeof(plate), d::Distribution{K}, len::Int; params = Parameters()) where {L <: ConstrainedSelection, K}

The convenience generate function provides an easy API to the Generate context. You can use this function on any of the matching signatures above - it will return the return value ret, a RecordSite instance specialized to the call, and the score/weight w computed with respect to the constraints sel.

Jaynes.SimulateContext — Type

mutable struct SimulateContext{T <: Trace} <: ExecutionContext
    tr::T
    visited::Visitor
    params::LearnableParameters
    SimulateContext(params) where T <: Trace = new{T}(Trace(), Visitor(), params)
end

SimulateContext is used to simulate traces without recording likelihood weights. SimulateContext can be instantiated with custom LearnableParameters instances, which is useful when used for gradient-based learning.

Inner constructors:

SimulateContext(params) = new{HierarchicalTrace}(Trace(), Visitor(), params)

Missing docstring.

Missing docstring for simulate. Check Documenter's build log for details.

Jaynes.ProposeContext — Type

mutable struct ProposeContext{T <: Trace} <: ExecutionContext
    tr::T
    weight::Float64
    score::Float64
    visited::Visitor
    params::LearnableParameters
end

ProposeContext is used to propose traces for inference algorithms which use custom proposals. ProposeContext instances can be passed sets of LearnableParameters to configure the propose with parameters which have been learned by differentiable programming.

Inner constructors:

ProposeContext(tr::T) where T <: Trace = new{T}(tr, 0.0, LearnableParameters())

Outer constructors:

Propose() = ProposeContext(Trace())

Jaynes.propose — Function

ret, g_cl, w = propose(fn::Function, args...)
ret, cs, w = propose(fn::typeof(rand), d::Distribution{K}) where K
ret, v_cl, w = propose(fn::typeof(markov), call::Function, len::Int, args...)
ret, v_cl, w = propose(fn::typeof(plate), call::Function, args::Vector)
ret, v_cl, w = propose(fn::typeof(plate), d::Distribution{K}, len::Int) where K

The convenience propose function provides an easy API to the Propose context. You can use this function on any of the matching signatures above - it will return the return value ret, a RecordSite instance specialized to the call, and the score/weight w.

Jaynes.ScoreContext — Type

mutable struct ScoreContext{P <: Parameters} <: ExecutionContext
    select::ConstrainedSelection
    weight::Float64
    params::P
end

The ScoreContext is used to score selections according to a model function. For computation in the ScoreContext to execute successfully, the select selection must provide a choice for every address visited in the model function, and the model function must allow the context to visit every constraints expressed in select.

Inner constructors:

function Score(obs::Vector{Tuple{K, P}}) where {P, K <: Union{Symbol, Pair}}
    c_sel = selection(obs)
    new{NoParameters}(c_sel, 0.0, Parameters())
end

Outer constructors:

ScoreContext(obs::K, params) where {K <: ConstrainedSelection} = new(obs, 0.0, params)
Score(obs::Vector) = ScoreContext(selection(obs))
Score(obs::ConstrainedSelection) = ScoreContext(obs, Parameters())
Score(obs::ConstrainedSelection, params) = ScoreContext(obs, params)

Missing docstring.

Missing docstring for score. Check Documenter's build log for details.

Missing docstring.

Missing docstring for UpdateContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for update. Check Documenter's build log for details.

Jaynes.RegenerateContext — Type

mutable struct RegenerateContext{T <: Trace, 
                                 L <: UnconstrainedSelection,
                                 P <: Parameters} <: ExecutionContext
    prev::T
    tr::T
    select::L
    weight::Float64
    score::Float64
    discard::T
    visited::Visitor
    params::P
end

Inner constructors:

function RegenerateContext(tr::T, sel::Vector{Address}) where T <: Trace
    un_sel = selection(sel)
    new{T, typeof(un_sel), NoParameters}(tr, Trace(), un_sel, 0.0, Trace(), Visitor(), Parameters())
end
function RegenerateContext(tr::T, sel::L) where {T <: Trace, L <: UnconstrainedSelection}
    new{T, L, NoParameters}(tr, Trace(), sel, 0.0, Trace(), Visitor(), Parameters())
end

Outer constructors:

Regenerate(tr::Trace, sel::Vector{Address}) = RegenerateContext(tr, sel)
Regenerate(tr::Trace, sel::UnconstrainedSelection) = RegenerateContext(tr, sel)

The RegenerateContext is used for MCMC algorithms, to propose new choices for addresses indicated by an UnconstrainedSelection in the select field.

Jaynes.regenerate — Function

ret, call_site = regenerate(sel::L, bbcs::HierarchicalCallSite, new_args...) where L <: UnconstrainedSelection
ret, call_site = regenerate(sel::L, bbcs::HierarchicalCallSite) where L <: UnconstrainedSelection

Users will likely interact with the RegenerateContext through the convenience method regenerate which requires that users provide an UnconstrainedSelection, an original call site, and possibly a set of new arguments to be used in the regeneration step. This context internally keeps track of the bookkeeping required to increment likelihood weights, as well as prune off parts of the trace which are invalid if a regenerated choice changes the shape of the trace (e.g. control flow).